# internal/server/openapi.yaml
openapi: 3.0.3
info:
  title: GeoVerse Map Server
  version: dev
  description: >-
    OGC API - Common, OGC API - Features and OGC API - Tiles surface of this
    service. This document intentionally does not cover the service's other
    endpoints (XYZ tile templates, WMTS, PMTiles archive Range downloads,
    spatial algorithms, MCP, /admin/* management) — those are this server's
    own extensions, not OGC API resources, and are documented in
    docs/DESIGN.md and README.md instead. See GET /conformance for the list
    of conformance classes this document is verified against.
servers:
  - url: /
paths:
...
  /collections/{collectionId}:
    get:
      operationId: getCollection
      summary: A feature collection
      parameters:
        - $ref: "#/components/parameters/collectionId"
      responses:
        "200": { description: Collection metadata. }
        "404": { description: Unknown collection. }
...
components:
  parameters:
    collectionId:
      name: collectionId
      in: path
      required: true
      schema: { type: string }
...